refactor(leaderboard): extract shared table controls and chart config
